requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
pandas>=1.2.0
matplotlib>=3.3.2
scikit-learn>=0.24.0
tqdm>=4.50.0
PyYAML>=5.4.0
pillow>=8.0.0
tensorboard>=2.4.0
lpips>=0.1.4
ftfy>=6.0.3
regex>=2020.10.28
clip @ git+https://github.com/openai/CLIP.git
einops>=0.3.0 